# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie REST API",
        "description": "A Laravel backend-only REST API wrapping TMDB, with JWT authentication, caching, and a personal favorites system.\n\n**Links:** [GitHub](https://github.com/codebyyosry) • [Medium](https://medium.com/@yosry.jobs)",
        "contact": {
            "name": "Yosry Badr",
            "url": "https://codebyyosry.github.io/#contact",
            "email": "yosry.jobs@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1": {
            "servers": [
                {
                    "url": "http://localhost:8001",
                    "description": "Local development server"
                }
            ]
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ahmed"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmed@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered, returns JWT"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and receive a JWT",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmed@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful, returns JWT"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the currently authenticated user",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Current user data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the current user (invalidate token)",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh the JWT token",
                "operationId": "30c20c52f5f72828d8c531fd7c9ea3e7",
                "responses": {
                    "200": {
                        "description": "New JWT issued"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "List the authenticated user's favorite movies",
                "operationId": "bc433e786ab5a0bec81c53c578e0b6aa",
                "responses": {
                    "200": {
                        "description": "List of favorites"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a movie to favorites",
                "operationId": "d426a8b38354581cb4082ac4bc62e22f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tmdb_movie_id"
                                ],
                                "properties": {
                                    "tmdb_movie_id": {
                                        "type": "integer",
                                        "example": 155
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Added to favorites"
                    },
                    "409": {
                        "description": "Movie already in favorites"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/{tmdbMovieId}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove a movie from favorites",
                "operationId": "810773231522c9464c8ebf3cb1a515ed",
                "parameters": [
                    {
                        "name": "tmdbMovieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 155
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed from favorites"
                    },
                    "404": {
                        "description": "Favorite not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get the list of movie genres",
                "operationId": "0ff78353b1c27ef8905e58f3cf31370a",
                "responses": {
                    "200": {
                        "description": "List of genres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 28
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Action"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/movies/popular": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a list of popular movies",
                "operationId": "f80b9fb4f87241954d3206e057760091",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of popular movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 155
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "The Dark Knight"
                                                    },
                                                    "overview": {
                                                        "type": "string",
                                                        "example": "Batman raises the stakes..."
                                                    },
                                                    "release_date": {
                                                        "type": "string",
                                                        "example": "2008-07-16"
                                                    },
                                                    "rating": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 8.5
                                                    },
                                                    "poster_url": {
                                                        "type": "string",
                                                        "example": "https://image.tmdb.org/t/p/w500/abc.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_pages": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "total_results": {
                                                    "type": "integer",
                                                    "example": 10000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/movies/search": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Search for movies by title",
                "operationId": "95e822a5784d90178255ea40c4b4fc3b",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search text",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "batman"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matching movies"
                    },
                    "422": {
                        "description": "Missing or invalid query parameter"
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get details for a single movie",
                "operationId": "0665e4e5eeb4738291d9733b6015218b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TMDB movie ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 155
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie details"
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Movies",
            "description": "Browse and search movies (public)"
        },
        {
            "name": "Genres",
            "description": "List movie genres (public)"
        },
        {
            "name": "Auth",
            "description": "Register, login, and manage authentication"
        },
        {
            "name": "Favorites",
            "description": "Authenticated users manage their favorite movies"
        }
    ]
}